refactor: relocate each component in its own directory
