+++ packages/ui/src/components/Composite/RegisterFlow/BasicInfoForm1/BasicInfoForm1.tsx
@@ -6,7 +6,7 @@ import { FolderPenIcon } from "@yamada-ui/lucide"
 import { FormControl, memo, noop, VStack } from "@yamada-ui/react"
 import type { FC } from "react"
 import { type SubmitHandler, useForm } from "react-hook-form"
-import { BasicInfoForm1Schema, type BasicInfoForm1Values } from "./schema"
+import { BasicInfoForm1Schema, type BasicInfoForm1Values } from "../schema"
 
 /**
  * Props for {@link BasicInfoForm1} component
