+++ packages/ui/src/components/Composite/RegisterFlow/BasicInfoForm1/index.ts
@@ -1,0 +1,1 @@
+export * from "./BasicInfoForm1"
