+++ packages/ui/src/components/Composite/RegisterFlow/BasicInfoForm2/index.ts
@@ -1,0 +1,1 @@
+export * from "./BasicInfoForm2"
