+++ packages/ui/src/components/Composite/RegisterFlow/UniversityInfoForm/index.ts
@@ -1,0 +1,1 @@
+export * from "./UniversityInfoForm"
